p4studio/requirements.txt
click==7.1.2
click-logging==1.0.1
cmakeast==0.0.18
PyYAML==5.3.1
jsonschema==2.6.0
